+++ bin/rialto-parachain/runtime/src/millau_messages.rs
@@ -62,10 +62,6 @@ pub type FromMillauMessageDispatch =
 /// Messages proof for Millau -> RialtoParachain messages.
 pub type FromMillauMessagesProof = messages::target::FromBridgedChainMessagesProof<bp_millau::Hash>;
 
-/// Messages delivery proof for RialtoParachain -> Millau messages.
-pub type ToMillauMessagesDeliveryProof =
-	messages::source::FromBridgedChainMessagesDeliveryProof<bp_millau::Hash>;
-
 /// Maximal outbound payload size of Rialto -> Millau messages.
 pub type ToMillauMaximalOutboundPayloadSize =
 	messages::source::FromThisChainMaximalOutboundPayloadSize<WithMillauMessageBridge>;
